Final version with option API vuejs

+++ src/modules/daybook/views/EntryView.vue
@@ -7,12 +7,14 @@
         <span class="mx-2 fs-4 fw-light">{{yearDay}}</span>
       </div>
       <div>
+
+        <input type="file" accept="image/jpg, image/png, image/jpeg" ref="imageSelector" v-show="false" @change="handleImage">
         <button v-if="this.entry.id" class="btn btn-danger mx-2" @click="removeEntry">
           Borrar
           <i class="fa fa-trash-alt"></i>
         </button>
 
-        <button class="btn btn-primary">
+        <button class="btn btn-primary" @click="imageButtonSelector">
           Subir foto
           <i class="fa fa-upload"></i>
         </button>
@@ -27,7 +29,14 @@
     <FabIcon icon="fa-save" @on:click="saveEntry" />
 
     <img
-      src="https://images.ctfassets.net/hrltx12pl8hq/a2hkMAaruSQ8haQZ4rBL9/8ff4a6f289b9ca3f4e6474f29793a74a/nature-image-for-website.jpg?fit=fill&w=480&h=320"
+      v-if="entry.picture && !loadImage"
+      :src="entry.picture"
+      alt="entry-pictre"
+      class="img-thumbnail"
+    />
+    <img
+      v-if="loadImage"
+      :src="loadImage"
       alt="entry-pictre"
       class="img-thumbnail"
     />
@@ -37,7 +46,9 @@
 <script>
 import { defineAsyncComponent } from "vue";
 import { mapActions, mapGetters } from "vuex";
+import Swal from 'sweetalert2'
 import getDayMonthYear from '../helpers/getDayMonthYear'
+import uploadImage from '../helpers/uploadImage'
 
 export default {
   props: {
@@ -49,6 +60,8 @@ export default {
   data(){
     return {
       entry: null,
+      loadImage: null,
+      file: null
     }
   },
   components: {
@@ -86,16 +99,57 @@ export default {
       this.entry = entry
     },
     async saveEntry(){
+      new Swal({
+        title: 'Espere por favor',
+        allowOutsideClick: false
+      })
+      Swal.showLoading()
+
+      this.entry.picture = await uploadImage( this.file )
+
       if(this.entry.id){
         await this.updateEntry(this.entry)
       } else {
         const newId = await this.createEntry(this.entry)
         this.$router.push({name: 'entry', params: {id: newId}})
       }
+      this.loadImage = null
+      Swal.fire('Guardado', 'Entrada registrada con éxtio', 'success')
     },
     async removeEntry(){
-      await this.deleteEntry(this.entry.id)
-      this.$router.push({name: 'no-entry'})
+      const { isConfirmed } = await Swal.fire({
+        title: '¿Está seguro?',
+        text: 'Una vez borrado no se puede recuperar',
+        showDenyButton: true,
+        confirmButtonText: 'Si, estoy seguro'
+      })
+      if( isConfirmed){
+        new Swal({
+          title: 'Espere por favor',
+          allowOutsideClick: false
+        })
+        Swal.showLoading()
+        await this.deleteEntry(this.entry.id)
+        this.$router.push({name: 'no-entry'})
+        Swal.fire('Eliminado','','success')
+      }
+    },
+    handleImage(event){
+      const file = event.target.files[0]
+      if(!file){
+        this.loadImage = null
+        this.file = null
+        return
+      }
+
+      this.file = file
+
+      const fr = new FileReader()
+      fr.onload = () => this.loadImage = fr.result
+      fr.readAsDataURL( file )
+    },
+    imageButtonSelector(){
+      this.$refs.imageSelector.click()
     }
   },
   created(){
